+++ crates/biome_configuration/src/linter/mod.rs
@@ -3,6 +3,7 @@ mod rules;
 
 pub use crate::linter::rules::Rules;
 use biome_analyze::options::RuleOptions;
+use biome_analyze::RuleFilter;
 use biome_deserialize::{Deserializable, StringSet};
 use biome_deserialize::{DeserializableValue, DeserializationDiagnostic, Merge, VisitableType};
 use biome_deserialize_macros::{Deserializable, Merge, Partial};
@@ -218,4 +219,80 @@ impl FromStr for RulePlainConfiguration {
 pub struct RuleWithOptions<T: Default> {
     pub level: RulePlainConfiguration,
     pub options: T,
+}
+
+#[derive(Clone, Copy, Debug, Eq, PartialEq, Hash)]
+pub enum RuleSelector {
+    Group(RuleGroup),
+    Rule(RuleGroup, &'static str),
+}
+
+impl From<RuleSelector> for RuleFilter<'static> {
+    fn from(value: RuleSelector) -> Self {
+        match value {
+            RuleSelector::Group(group) => RuleFilter::Group(group.as_str()),
+            RuleSelector::Rule(group, name) => RuleFilter::Rule(group.as_str(), name),
+        }
+    }
+}
+
+impl FromStr for RuleSelector {
+    type Err = &'static str;
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        if let Some((group_name, rule_name)) = s.split_once('/') {
+            let group = RuleGroup::from_str(group_name)?;
+            if let Some(rule_name) = Rules::has_rule(group, rule_name) {
+                Ok(RuleSelector::Rule(group, rule_name))
+            } else {
+                Err("This rule doesn't exist.")
+            }
+        } else {
+            match RuleGroup::from_str(s) {
+                Ok(group) => Ok(RuleSelector::Group(group)),
+                Err(_) => Err(
+                    "This group doesn't exist. Use the syntax `<group>/<rule>` to specify a rule.",
+                ),
+            }
+        }
+    }
+}
+
+impl serde::Serialize for RuleSelector {
+    fn serialize<S: serde::Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
+        match self {
+            RuleSelector::Group(group) => serializer.serialize_str(group.as_str()),
+            RuleSelector::Rule(group, rule_name) => {
+                let group_name = group.as_str();
+                serializer.serialize_str(&format!("{group_name}/{rule_name}"))
+            }
+        }
+    }
+}
+
+impl<'de> serde::Deserialize<'de> for RuleSelector {
+    fn deserialize<D: serde::Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
+        struct Visitor;
+        impl<'de> serde::de::Visitor<'de> for Visitor {
+            type Value = RuleSelector;
+            fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
+                formatter.write_str("<group>/<ruyle_name>")
+            }
+            fn visit_str<E: serde::de::Error>(self, v: &str) -> Result<Self::Value, E> {
+                match RuleSelector::from_str(v) {
+                    Ok(result) => Ok(result),
+                    Err(error) => Err(serde::de::Error::custom(error)),
+                }
+            }
+        }
+        deserializer.deserialize_str(Visitor)
+    }
+}
+
+impl schemars::JsonSchema for RuleSelector {
+    fn schema_name() -> String {
+        "RuleCode".to_string()
+    }
+    fn json_schema(gen: &mut schemars::gen::SchemaGenerator) -> schemars::schema::Schema {
+        String::json_schema(gen)
+    }
 }