+++ CHANGELOG.md
@@ -15,6 +15,43 @@ our [guidelines for writing a good changelog entry](https://github.com/biomejs/b
 
 ### CLI
 
+#### New features
+
+- Add a new option `--rule` to the command `biome lint` ([#58](https://github.com/biomejs/biome/issues/58)).
+
+  This new option allows you to execute a single rule or a rule group.
+  This option is convenient to test a rule or apply the code fixes of a single rule.
+
+  For example, you can execute the `style/useNamingConvention` rule on the working directory:
+
+  ```shell
+  biome lint --rule=style/useNamingConvention ./
+  ```
+
+  If the rule has a code action (autofix), you can use `--apply` to apply the fix:
+
+  ```shell
+  biome lint --rule=style/useNamingConvention --apply ./
+  ```
+
+  The option takes the rule options in the Biome configuration file into account.
+  Only, the severity level of the rule is overridden by its default value,
+  i.e. `error` for a recommended rule or `warn` otherwise.
+
+  You can also run a group of rules:
+
+  ```shell
+  biome lint --rule=suspicous src/main.js
+  ```
+
+  In this case, the severity level of a rule is not overridden.
+  Thus, disabled rules stay disabled.
+  To ensure that the group is run, the `recommended` field of the group is turned on.
+
+  The option is compatible with other options such as `--apply`, `--apply-unsafe` and `--reporter`.
+
+  Contributed by @Conaclos
+
 #### Enhancements
 
 - Biome now executes commands (lint, format, check and ci) on the working directory by default. [#2266](https://github.com/biomejs/biome/issues/2266) Contributed by @unvalley